+++ src/providers/prisma-provider.ts
@@ -1,5 +1,6 @@
+import { Injectable } from '@nestjs/common';
 import { PrismaClient } from '@prisma/client';
-
+@Injectable()
 export class PrismaProvider {
   readonly client: PrismaClient;
   constructor() {